Move return type inference code to its own file
---
