+++ frontend/lib/resolution/Makefile.include
@@ -33,6 +33,7 @@ FRONTEND_RESOLUTION_SRCS =                          \
            prims.cpp                                \
            resolution-queries.cpp                   \
            resolution-types.cpp                     \
+           return-type-inference.cpp                \
            scope-queries.cpp                        \
            scope-types.cpp                          \
            split-init.cpp                           \
